+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -32,6 +32,18 @@ public class BoxTestGenerated extends AbstractBoxTest {
     }
 
     @Test
+    @TestMetadata("ContributedBindingContainerExclusions.kt")
+    public void testContributedBindingContainerExclusions() {
+      runTest("compiler-tests/src/test/data/box/aggregation/ContributedBindingContainerExclusions.kt");
+    }
+
+    @Test
+    @TestMetadata("ContributedBindingContainerReplacements.kt")
+    public void testContributedBindingContainerReplacements() {
+      runTest("compiler-tests/src/test/data/box/aggregation/ContributedBindingContainerReplacements.kt");
+    }
+
+    @Test
     @TestMetadata("ContributesMultibindingInteropAnnotationsAddBindingToSetOrMapWithMapKey.kt")
     public void testContributesMultibindingInteropAnnotationsAddBindingToSetOrMapWithMapKey() {
       runTest("compiler-tests/src/test/data/box/aggregation/ContributesMultibindingInteropAnnotationsAddBindingToSetOrMapWithMapKey.kt");
@@ -212,6 +224,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       @TestMetadata("BindingContainerViaAnnotationTransitive.kt")
       public void testBindingContainerViaAnnotationTransitive() {
         runTest("compiler-tests/src/test/data/box/dependencygraph/bindingcontainers/BindingContainerViaAnnotationTransitive.kt");
+      }
+
+      @Test
+      @TestMetadata("BindingContainerViaContributesTo.kt")
+      public void testBindingContainerViaContributesTo() {
+        runTest("compiler-tests/src/test/data/box/dependencygraph/bindingcontainers/BindingContainerViaContributesTo.kt");
       }
 
       @Test