+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/DependencyGraphTransformer.kt
@@ -83,7 +83,8 @@ internal class DependencyGraphTransformer(
   // Keyed by the source declaration
   private val processedMetroDependencyGraphsByClass = mutableMapOf<ClassId, IrClass>()
 
-  private val dependencyGraphNodeCache = DependencyGraphNodeCache(this, bindingContainerTransformer)
+  private val dependencyGraphNodeCache =
+    DependencyGraphNodeCache(this, contributionData, bindingContainerTransformer)
 
   override fun visitCall(expression: IrCall): IrExpression {
     return CreateGraphTransformer.visitCall(expression, metroContext)